+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Summary of release changes for Version 1 - CentOS-6
 - Updates container naming conventions and readability of `Makefile`.
 - Updates `docker logs` output example in README document.
 - Updates README instructions following review.
+- Updates default HEALTHCHECK interval to 1 second from 0.5.
 - Replaces awk with native bash regex when testing sudo user's have `NOPASSWD:ALL`.
 - Fixes bootstrap errors regarding readonly `PASSWORD_LENGTH`.
 - Fixes issue with redacted password when using `SSH_PASSWORD_AUTHENTICATION` in combination with `SSH_USER_FORCE_SFTP`.
@@ -22,6 +23,7 @@ Summary of release changes for Version 1 - CentOS-6
 - Adds error messages to healthcheck script and includes supervisord check.
 - Adds `__docker_logs_match` function to test cases to work-around output delays on CI service's host.
 - Adds port incrementation to Makefile's run template for container names with an instance suffix.
+- Adds consideration for event lag into test cases for unhealthy health_status events.
 - Removes use of `/etc/services-config` paths.
 - Removes fleet `--manager` option in the `scmi` installer.
 - Removes X-Fleet section from etcd register template unit-file.